MMCA-5406 Handle absent responseDetail for all Sub09 calls and log appropriately
